automatically backed up by learn

+++ index.js
@@ -27,9 +27,9 @@ function makeInt(n){
 }
 
 function preserveDecimal(n){
-  return parseFloat('2.222', 10)
+  return parseInt('2.222', 10)
 }
 
 function preserveDecimal(n){
-  return parseInt('2.222', 10)
+  return parseFloat('2.222', 10)
 }
